Always download the latest ModTheSpire.jar (provide optional override)

+++ bin/download_dependencies.sh
@@ -1,4 +1,5 @@
-MOD_THE_SPIRE_ZIP_URL=https://github.com/kiooeht/ModTheSpire/releases/download/v2.9.1/ModTheSpire.zip
+MOD_THE_SPIRE_LATEST_RELEASE_URL=https://github.com/kiooeht/ModTheSpire/releases/latest
+#MOD_THE_SPIRE_ZIP_URL=https://github.com/kiooeht/ModTheSpire/releases/download/v2.9.1/ModTheSpire.zip
 BASE_MOD_JAR_URL=https://github.com/daviscook477/BaseMod/releases/download/v2.18.0/BaseMod.jar
 DESKTOP_JAR_LOCAL_PATH="$HOME/Library/Application Support/Steam/steamapps/common/SlayTheSpire/SlayTheSpire.app/Contents/Resources/desktop-1.0.jar"
 
@@ -9,11 +10,20 @@ fi
 getModTheSpireJar() {
   if [ -f ./ModTheSpire.jar ]; then
     echo "ModTheSpire.jar already present - skipping."
-  else
-    echo "Downloading ModTheSpire.jar"
+    return
+  fi
+
+  if [ -z $MOD_THE_SPIRE_ZIP_URL ]; then
+    MOD_THE_SPIRE_ZIP_URL=`curl $MOD_THE_SPIRE_LATEST_RELEASE_URL 2> /dev/null | sed -E "s/.*href=\"([^\"]*)\".*/\1\/ModTheSpire.zip/g; s/tag/download/g"`
+  fi
+
+  if [ ! -z $MOD_THE_SPIRE_ZIP_URL ]; then
+    echo "Downloading ModTheSpire.jar from $MOD_THE_SPIRE_ZIP_URL"
     curl -L $MOD_THE_SPIRE_ZIP_URL > ModTheSpire.zip
     unzip -d . ModTheSpire.zip ModTheSpire.jar
     rm ModTheSpire.zip
+  else
+    echo "Could not download ModTheSpire.jar - try downloading it manually from $MOD_THE_SPIRE_LATEST_RELEASE_URL"
   fi
 }
